+++ src/main/java/org/programmers/staybb/Controller/User/UserController.java
@@ -3,10 +3,9 @@ package org.programmers.staybb.Controller.User;
 import javassist.NotFoundException;
 import javax.validation.Valid;
 import org.programmers.staybb.dto.user.UserRequest;
-import org.programmers.staybb.global.ApiResponse;
-import org.programmers.staybb.global.ApiUtils;
-import org.programmers.staybb.service.user.UserService;
-import org.springframework.beans.factory.annotation.Value;
+import org.programmers.staybb.global.response.ApiResponse;
+import org.programmers.staybb.global.response.ApiUtils;
+import org.programmers.staybb.service.UserService;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
