+++ src/main/java/org/programmers/staybb/global/exception/GlobalException.java
@@ -1,10 +1,10 @@
-package org.programmers.staybb.global;
+package org.programmers.staybb.global.exception;
 
 import javassist.NotFoundException;
+import org.programmers.staybb.global.response.ApiResponse;
+import org.programmers.staybb.global.response.ApiUtils;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.client.HttpClientErrorException.NotFound;
 
 @RestControllerAdvice
 public class GlobalException {
@@ -13,5 +13,7 @@ public class GlobalException {
     public ApiResponse notFoundHandler(NotFoundException e) {
         return ApiUtils.error(e.getMessage(), 404);
     }
+
+
 
 }
