+++ src/main/java/org/programmers/staybb/global/validation/DateValid.java
@@ -1,6 +1,5 @@
-package org.programmers.staybb;
+package org.programmers.staybb.global.validation;
 
-import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 import javax.validation.Constraint;
