+++ src/main/java/org/programmers/staybb/global/validation/DateValidator.java
@@ -1,4 +1,4 @@
-package org.programmers.staybb;
+package org.programmers.staybb.global.validation;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
