+++ src/main/java/org/programmers/staybb/repository/UserRepository.java
@@ -2,7 +2,7 @@ package org.programmers.staybb.repository;
 
 import org.programmers.staybb.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
+
 
 public interface UserRepository extends JpaRepository<User, Long> {
